+++ backend/celery_start.sh
@@ -2,7 +2,7 @@
 set -e
 
 # Source Redis URL construction
-. /app/construct-redis-url.sh
+. /app/construct-connection-strings.sh
 
 set -o errexit
 set -o nounset